Refactor the route registry slightly
